add:docstrings

+++ 2-do_deploy_web_static.py
@@ -23,11 +23,13 @@ def do_deploy(archive_path):
 
     if put(archive_path, "/tmp/{}".format(file)).failed is True:
         return False
+    if run("mkdir -p /data/web_static/releases/{}/".
+           format(name)).failed is True:
+        return False
     if run("rm -rf /data/web_static/releases/{}/".
            format(name)).failed is True:
         return False
-    if run("mkdir -p /data/web_static/releases/{}/".
-           format(name)).failed is True:
+    if run("rm -rf /data/web_static/current").failed is True:
         return False
     if run("tar -xzf /tmp/{} -C /data/web_static/releases/{}/".
            format(file, name)).failed is True:
@@ -39,8 +41,6 @@ def do_deploy(archive_path):
         return False
     if run("rm -rf /data/web_static/releases/{}/web_static".
            format(name)).failed is True:
-        return False
-    if run("rm -rf /data/web_static/current").failed is True:
         return False
     if run("ln -s /data/web_static/releases/{}/ /data/web_static/current".
            format(name)).failed is True:
